refactor: reuse models from core
