+++ packages/core/src/index.ts
@@ -7,3 +7,10 @@ export type {
 export { executeRequest } from './http-client.js';
 export { getAiCoreDestination } from './context.js';
 export { OpenApiRequestBuilder } from './openapi-request-builder.js';
+export {
+  AzureOpenAiChatModel,
+  AzureOpenAiEmbeddingModel,
+  GcpVertexAiChatModel,
+  AwsBedrockChatModel,
+  AicoreOpensourceChatModel
+} from './model-types.js';
