+++ packages/core/src/internal.ts
@@ -1,3 +1,4 @@
 export * from './context.js';
 export * from './http-client.js';
 export * from './openapi-request-builder.js';
+export * from './model-types.js';
