+++ packages/core/src/model-types.ts
@@ -1,0 +1,44 @@
+type LiteralUnion<T extends U, U = string> = T | (U & Record<never, never>);
+
+/**
+ * Azure OpenAI models for chat completion.
+ */
+export type AzureOpenAiChatModel = LiteralUnion<
+  | 'gpt-4o-mini'
+  | 'gpt-4o'
+  | 'gpt-4'
+  | 'gpt-4-32k'
+  | 'gpt-35-turbo'
+  | 'gpt-35-turbo-16k'
+>;
+
+/**
+ * Azure OpenAI models for embedding.
+ */
+export type AzureOpenAiEmbeddingModel = LiteralUnion<
+  'text-embedding-ada-002' | 'text-embedding-3-small' | 'text-embedding-3-large'
+>;
+
+/**
+ * GCP Vertex AI models for chat completion.
+ */
+export type GcpVertexAiChatModel = LiteralUnion<
+  'gemini-1.0-pro' | 'gemini-1.5-pro' | 'gemini-1.5-flash'
+>;
+
+/**
+ * AWS Bedrock models for chat completion.
+ */
+export type AwsBedrockChatModel = LiteralUnion<
+  | 'anthropic--claude-3-haiku'
+  | 'anthropic--claude-3-opus'
+  | 'anthropic--claude-3-sonnet'
+  | 'anthropic--claude-3.5-sonnet'
+>;
+
+/**
+ * Aicore Opensource models for chat completion.
+ */
+export type AicoreOpensourceChatModel = LiteralUnion<
+  'mistralai--mixtral-8x7b-instruct-v01' | 'meta--llama3-70b-instruct'
+>;
