+++ packages/foundation-models/src/openai/model-types.ts
@@ -1,0 +1,14 @@
+import type {
+  AzureOpenAiChatModel as CoreAzureOpenAiChatModel,
+  AzureOpenAiEmbeddingModel as CoreAzureOpenAiEmbeddingModel
+} from '@sap-ai-sdk/core';
+
+/**
+ * Azure OpenAI models for chat completion.
+ */
+export type AzureOpenAiChatModel = CoreAzureOpenAiChatModel;
+
+/**
+ * Azure OpenAI models for embedding.
+ */
+export type AzureOpenAiEmbeddingModel = CoreAzureOpenAiEmbeddingModel;
