+++ packages/foundation-models/src/openai/openai-chat-client.ts
@@ -3,7 +3,7 @@ import {
   getDeploymentId,
   type ModelDeployment
 } from '@sap-ai-sdk/ai-api/internal.js';
-import type { AzureOpenAiChatModel } from '../model-types.js';
+import type { AzureOpenAiChatModel } from './model-types.js';
 import type { OpenAiChatCompletionParameters } from './openai-types.js';
 import { OpenAiChatCompletionResponse } from './openai-response.js';
