+++ packages/orchestration/src/index.ts
@@ -40,10 +40,4 @@ export { azureContentFilter } from './orchestration-filter-utility.js';
 
 export { OrchestrationResponse } from './orchestration-response.js';
 
-export type {
-  AzureOpenAiChatModel,
-  AicoreOpensourceChatModel,
-  GcpVertexAiChatModel,
-  AwsBedrockChatModel,
-  ChatModel
-} from './model-types.js';
+export type { ChatModel } from './model-types.js';
