+++ packages/orchestration/src/model-types.ts
@@ -1,42 +1,17 @@
-type LiteralUnion<T extends U, U = string> = T | (U & Record<never, never>);
-
-/**
- * Azure OpenAI models for chat completion.
- */
-export type AzureOpenAiChatModel = LiteralUnion<
-  'gpt-4o' | 'gpt-4' | 'gpt-4-32k' | 'gpt-35-turbo' | 'gpt-35-turbo-16k'
->;
-
-/**
- * GCP Vertex AI models for chat completion.
- */
-export type GcpVertexAiChatModel = LiteralUnion<
-  'gemini-1.0-pro' | 'gemini-1.5-pro' | 'gemini-1.5-flash'
->;
-
-/**
- * AWS Bedrock models for chat completion.
- */
-export type AwsBedrockChatModel = LiteralUnion<
-  | 'anthropic--claude-3-haiku'
-  | 'anthropic--claude-3-opus'
-  | 'anthropic--claude-3-sonnet'
-  | 'anthropic--claude-3.5-sonnet'
->;
-
-/**
- * Aicore Opensource models for chat completion.
- */
-export type AicoreOpensourceChatModel = LiteralUnion<
-  'mistralai--mixtral-8x7b-instruct-v01' | 'meta--llama3-70b-instruct'
->;
+import {
+  AicoreOpensourceChatModel,
+  AwsBedrockChatModel,
+  AzureOpenAiChatModel,
+  GcpVertexAiChatModel
+} from '@sap-ai-sdk/core';
 
 /**
  * Supported chat models for orchestration.
  */
-export type ChatModel = LiteralUnion<
+export type ChatModel = Exclude<
   | AzureOpenAiChatModel
   | GcpVertexAiChatModel
   | AwsBedrockChatModel
-  | AicoreOpensourceChatModel
+  | AicoreOpensourceChatModel,
+  'gpt-4o-mini'
 >;
